requirements.txt
altair==5.5.0
numpy==2.0.1
opencv-python-headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
streamlit==1.41.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
ultralytics==8.3.56
openpyxl